Bug fixes and improvements

+++ matlab/Micromet/create_TAB_ProjectFolders.m
@@ -15,10 +15,16 @@ function create_TAB_ProjectFolders(projectPath,siteID)
 %
 %
 % Zoran Nesic           File created:       Aug 23, 2024
-%                       Last modification:  Aug 23, 2024
+%                       Last modification:  Sep 13, 2024
 
 % Revisions
 %
+% Sep 13, 2024 (Zoran)
+%   - Added a few warning dialogs that can prevent accidental overwriting
+%     of the local ini files.
+% Sep 12, 2024 (Zoran)
+%   - added repo for include files.
+%   (https://github.com/CANFLUX/TAB_include_files.git)
 
 if ~exist('projectPath','var')
     error('Missing Project Path.')
@@ -36,12 +42,10 @@ end
 if exist(fullfile(projectPath,'Database'),'dir') ...
    || exist(fullfile(projectPath,'Sites'),'dir') ...
    || exist(fullfile(projectPath,'Matlab'),'dir')
-    fig = uifigure;
-    selection = uiconfirm(fig, ...
-    "Project folders already exist. Do you want to overwrite them?","Confirm Overwrite", ...
-    "Options",["Overwrite","Cancel"], ...
-    "Icon","warning",'CloseFcn',@(h,e) close(fig));
-    if strcmpi(selection,'Cancel')
+     selection = questdlg('Project folders already exist! Do you want to overwrite them?', ...
+                     'Last warning!', ...
+                     'Yes', 'No','No');
+    if strcmpi(selection,'No')
         return
     end
 end
@@ -58,18 +62,69 @@ end
 tmpPath = fullfile(projectPath,'Database','Calculation_Procedures','AmeriFlux');
 if ~exist(tmpPath,'dir')
     mkdir(tmpPath);
+    url1 = 'https://github.com/CANFLUX/TAB_Ameriflux_variables.git';
+    gitclone(url1,tmpPath,Depth=1);
 else
     % For repo to be downloaded the folder needs to be empty so first remove it
-    rmdir(tmpPath,"s");
-    mkdir(tmpPath);
+    ButtonName = questdlg('AmeriFlux repo folder already exists. If you press OK it will be deleted. Your personalized files will be lost! Do you want to proceed?', ...
+                         'Deleting local files', ...
+                         'Yes', 'No','No');
+    if strcmpi(ButtonName,'Yes')
+        ButtonName = questdlg('Are you sure!', ...
+                     'Last warning!', ...
+                     'Yes', 'No','No');
+        if strcmpi(ButtonName,'Yes')
+            fprintf('Overwriting AmerFlux local copy.\n\n')
+            rmdir(tmpPath,"s");
+            mkdir(tmpPath);
+            url1 = 'https://github.com/CANFLUX/TAB_Ameriflux_variables.git';
+            gitclone(url1,tmpPath,Depth=1);            
+        else
+            fprintf('Skipping AmerFlux repo cloning.\n\n')
+        end
+    else
+        fprintf('Skipping AmerFlux repo cloning.\n\n')
+    end
 end
-url1 = 'https://github.com/CANFLUX/TAB_Ameriflux_variables.git';
-gitclone(url1,tmpPath,Depth=1);
 
 tmpPath = fullfile(projectPath,'Database','Calculation_Procedures','TraceAnalysis_ini');
 if ~exist(tmpPath,'dir')
     mkdir(tmpPath);
+    url1 = 'https://github.com/CANFLUX/TAB_include_files.git';
+    gitclone(url1,tmpPath,Depth=1);
+else
+    % For repo to be downloaded the folder needs to be empty so first remove it
+    ButtonName = questdlg('TraceAnalisis_ini folder already exists. If you press OK it will be deleted. Your personalized INI files will be lost! Do you want to proceed?', ...
+                         'Deleting local files', ...
+                         'Yes', 'No','No');
+    if strcmpi(ButtonName,'Yes')
+        ButtonName = questdlg('Are you sure!', ...
+                     'Last warning!', ...
+                     'Yes', 'No','No');
+        if strcmpi(ButtonName,'Yes')
+            fprintf('Overwriting TraceAnalisis_ini folder.\n\n')
+            rmdir(tmpPath,"s");
+            mkdir(tmpPath);
+            url1 = 'https://github.com/CANFLUX/TAB_include_files.git';
+            gitclone(url1,tmpPath,Depth=1);           
+        else
+            fprintf('Skipping TraceAnalisis_ini folder repo cloning.\n\n')
+        end
+    else
+        fprintf('Skipping TraceAnalisis_ini folder repo cloning.\n\n')
+    end
 end
+
+% tmpPath = fullfile(projectPath,'Database','Calculation_Procedures','TraceAnalysis_ini');
+% if ~exist(tmpPath,'dir')
+%     mkdir(tmpPath);
+% else
+%     % For repo to be downloaded the folder needs to be empty so first remove it
+%     rmdir(tmpPath,"s");
+%     mkdir(tmpPath);    
+% end
+% url1 = 'https://github.com/CANFLUX/TAB_include_files.git';
+% gitclone(url1,tmpPath,Depth=1);
 
 tmpPath = fullfile(projectPath,'Database','Calculation_Procedures','TraceAnalysis_ini',siteID);
 if ~exist(tmpPath,'dir')
